requirements-polymarket.txt
requests
py-clob-client
python-dotenv
